[refactor]: Separate component in signup page
