+++ src/api/axios.ts
@@ -14,4 +14,5 @@ axios.interceptors.request.use(
     return Promise.reject(error);
   },
 );
+
 export default axiosInstance;
