+++ src/components/common/Header.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
+import { NavLink } from 'react-router-dom';
 
 const Header = () => {
   return (
     <header>
-      <img
-        src="src/assets/icon/icon.svg"
-        alt="icon"
-        style={{ width: 48, height: 48 }}
-      />
+      <NavLink to="/">
+        <img
+          src="src/assets/icon/icon.svg"
+          alt="icon"
+          style={{ width: 48, height: 48 }}
+        />
+      </NavLink>
     </header>
   );
 };
