+++ src/stores/store.ts
@@ -1,5 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import authReducer from '#/stores/auth/authSlice';
+
+import authReducer from '#/stores/auth/slice';
 
 export const store = configureStore({
   reducer: {
